+++ client/coral-framework/actions/auth.js
@@ -137,3 +137,8 @@ export const checkLogin = () => dispatch => {
     })
     .catch(error => dispatch(checkLoginFailure(error)));
 };
+
+export const requestConfirmEmail = () => dispatch => {
+
+};
+
