+++ client/coral-sign-in/translations.js
@@ -1,6 +1,8 @@
 export default {
   en: {
     'signIn': {
+      emailConfirmCTA: 'Please verify your email address. We sent an email to {0} for verification.',
+      requestNewConfirmEmail: 'Request another email',
       notYou: 'Not you?',
       loggedInAs: 'Logged in as',
       facebookSignIn: 'Sign in with Facebook',
@@ -28,6 +30,8 @@ export default {
   },
   es: {
     'signIn': {
+      emailConfirmCTA: 'Por favor verifique su correo electronico. Le enviamos un correo a {0} para verificar.',
+      requestNewConfirmEmail: 'Enviar otro correo',
       notYou: 'No eres tu?',
       loggedInAs: 'Entraste como',
       facebookSignIn: 'Entrar con Facebook',
