+++ client/coral-ui/components/FormField.js
@@ -1,5 +1,5 @@
-import React from 'react';
-import styles from './styles.css';
+import React, {PropTypes} from 'react';
+import styles from './FormField.css';
 
 const FormField = ({className, showErrors = false, errorMsg, label, ...props}) => (
   <div className={`${styles.formField} ${className ? className : ''}`}>
@@ -14,5 +14,13 @@ const FormField = ({className, showErrors = false, errorMsg, label, ...props}) =
     {showErrors && errorMsg && <span className={styles.errorMsg}><span className={styles.attention}>!</span>{errorMsg}</span>}
   </div>
 );
+
+FormField.propTypes = {
+  label: PropTypes.string,
+  value: PropTypes.string,
+  onChange: PropTypes.func,
+  errorMsg: PropTypes.string,
+  type: PropTypes.string
+};
 
 export default FormField;
